libpebble3: Compose-free public API (PebbleBitmap), extract Contacts, relocate fakes
Decouple the published API surface from Compose so external consumers don't need
AndroidX/Compose, and tidy the surface for standalone consumption:

- Introduce image/PebbleBitmap.kt: a pure-Kotlin ARGB8888 bitmap. Replace
  ImageBitmap with PebbleBitmap in every published bitmap-returning API
  (Screenshot.takeScreenshot, NotificationApps.getAppIcon, Contacts.getContactImage)
  and all of their implementations. Add util conversion helpers
  (PebbleBitmap.toImageBitmap / Bitmap.toPebbleBitmap); app-side call sites convert
  at the point of rendering. Drop the @Stable Compose annotations from the interfaces.
- Extract the Contacts interface out of LibPebble.kt into its own contacts/Contacts.kt.
- Move FakeLibPebble from connection/ to a dedicated fake/ package and split the fake
  ConnectedPebbleDevice out into connection/FakeConnectedDevice.kt.

Entire-Checkpoint: 7694c2603501
