+++ libpebble3/src/androidMain/kotlin/io/rebble/libpebblecommon/util/BitmapUtil.android.kt
@@ -4,6 +4,14 @@ import android.graphics.Bitmap
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asImageBitmap
 import co.touchlab.kermit.Logger
+import io.rebble.libpebblecommon.image.PebbleBitmap
+
+fun Bitmap.toPebbleBitmap(): PebbleBitmap {
+    val argb = if (config == Bitmap.Config.ARGB_8888) this else copy(Bitmap.Config.ARGB_8888, false)
+    val pixels = IntArray(argb.width * argb.height)
+    argb.getPixels(pixels, 0, argb.width, 0, 0, argb.width, argb.height)
+    return PebbleBitmap(width = argb.width, height = argb.height, pixels = pixels)
+}
 
 actual fun createImageBitmapFromPixelArray(
     pixels: IntArray,
