+++ libpebble3/src/commonMain/kotlin/io/rebble/libpebblecommon/image/PebbleBitmap.kt
@@ -1,0 +1,20 @@
+package io.rebble.libpebblecommon.image
+
+class PebbleBitmap(
+    val width: Int,
+    val height: Int,
+    val pixels: IntArray,
+) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is PebbleBitmap) return false
+        return width == other.width && height == other.height && pixels.contentEquals(other.pixels)
+    }
+
+    override fun hashCode(): Int {
+        var result = width
+        result = 31 * result + height
+        result = 31 * result + pixels.contentHashCode()
+        return result
+    }
+}
