+++ libpebble3/src/commonMain/kotlin/io/rebble/libpebblecommon/notification/PlatformNotificationListener.kt
@@ -1,6 +1,5 @@
 package io.rebble.libpebblecommon.notification
 
-import androidx.compose.ui.graphics.ImageBitmap
 import io.rebble.libpebblecommon.connection.AppContext
 import io.rebble.libpebblecommon.connection.LibPebble
 import io.rebble.libpebblecommon.connection.NotificationApps
@@ -17,6 +16,7 @@ import io.rebble.libpebblecommon.database.entity.MuteState
 import io.rebble.libpebblecommon.database.entity.NotificationEntity
 import io.rebble.libpebblecommon.database.entity.VibePatternEntity
 import io.rebble.libpebblecommon.di.LibPebbleCoroutineScope
+import io.rebble.libpebblecommon.image.PebbleBitmap
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
 import kotlinx.coroutines.flow.Flow
@@ -142,7 +142,7 @@ class NotificationApi(
         }
     }
 
-    override suspend fun getAppIcon(packageName: String): ImageBitmap? {
+    override suspend fun getAppIcon(packageName: String): PebbleBitmap? {
         return withContext(Dispatchers.IO) {
             iconFor(packageName, appContext)
         }
@@ -180,4 +180,4 @@ class NotificationApi(
     }
 }
 
-expect fun iconFor(packageName: String, appContext: AppContext): ImageBitmap?
+expect fun iconFor(packageName: String, appContext: AppContext): PebbleBitmap?
