+++ libpebble3/src/jvmMain/kotlin/io/rebble/libpebblecommon/notification/PlatformNotificationListener.jvm.kt
@@ -1,6 +1,6 @@
 package io.rebble.libpebblecommon.notification
 
-import androidx.compose.ui.graphics.ImageBitmap
 import io.rebble.libpebblecommon.connection.AppContext
+import io.rebble.libpebblecommon.image.PebbleBitmap
 
-actual fun iconFor(packageName: String, appContext: AppContext): ImageBitmap? = null
+actual fun iconFor(packageName: String, appContext: AppContext): PebbleBitmap? = null
